+++ Core/Utility.lua
@@ -292,3 +292,9 @@ function MS:FetchMythicPlusInfo()
     end)
 end
 
+MS.AbbrRaidTitles = {
+    ["Vault of the Incarnates"] = "VOTI",               -- DF Season 1
+    ["Aberrus, the Shadowed Crucible"] = "ATSC",        -- DF Season 2
+    ["Amirdrassil, the Dream's Hope"] = "AOTD",         -- DF Season 3
+}
+